fix comments

+++ be/src/storage/virtual_column_utils.h
@@ -29,8 +29,8 @@ StatusOr<TabletSchemaCSPtr> extend_schema_by_virtual_columns(const TabletSchemaC
 class VirtualColumnFactory {
 public:
     struct Options {
-        int32_t tablet_id;
-        int32_t segment_id;
+        int64_t tablet_id;
+        int64_t segment_id;
         int64_t num_rows;
     };
     static StatusOr<ColumnIterator*> create_virtual_column_iterator(const Options& options,
